// apps/stripe/generated/app-webhooks-types/transaction-initialize-session.ts
export type Pspreference1 = string;
/**
 * Decimal amount of the processed action
 */
export type Amount1 = number | string;
/**
 * Time of the action in ISO 8601 format
 */
export type Time1 = string;
/**
 * External url with action details
 */
export type Externalurl1 = string;
/**
 * Message related to the action. The maximum length is 512 characters; any text exceeding this limit will be truncated
 */
export type Message1 = string;
export type Actions1 = ("CHARGE" | "REFUND" | "CANCEL")[] | null;
/**
 * Result of the action
 */
export type Result1 = "AUTHORIZATION_FAILURE" | "CHARGE_FAILURE";
/**
 * The JSON data that will be returned to storefront
 */
export type JsonValue1 =
  | string
  | number
  | boolean
  | {
      [k: string]: unknown;
    }
  | unknown[]
  | null;
/**
 * Details of the payment method used for the transaction.
 */
export type Paymentmethoddetails1 = (OtherPaymentMethodDetails | CardPaymentMethodDetails) | null;
/**
 * Psp reference received from payment provider.
 */
export type Pspreference2 = string;
/**
 * Decimal amount of the processed action
 */
export type Amount2 = number | string;
/**
 * Time of the action in ISO 8601 format
 */
export type Time2 = string;
/**
 * External url with action details
 */
export type Externalurl2 = string;
/**
 * Message related to the action. The maximum length is 512 characters; any text exceeding this limit will be truncated
 */
export type Message2 = string;
export type Actions2 = ("CHARGE" | "REFUND" | "CANCEL")[] | null;
/**
 * Result of the action
 */
export type Result2 = "AUTHORIZATION_ACTION_REQUIRED" | "CHARGE_ACTION_REQUIRED";
/**
 * The JSON data that will be returned to storefront
 */
export type JsonValue2 =
  | string
  | number
  | boolean
  | {
      [k: string]: unknown;
    }
  | unknown[]
  | null;
/**
 * Details of the payment method used for the transaction.
 */
export type Paymentmethoddetails2 = (OtherPaymentMethodDetails | CardPaymentMethodDetails) | null;

export interface TransactionSessionSuccess {
  pspReference: Pspreference;
  amount?: Amount;
  time?: Time;
  externalUrl?: Externalurl;
  message?: Message;
  actions?: Actions;
  result: Result;
  data?: JsonValue;
  paymentMethodDetails?: Paymentmethoddetails;
}
export interface OtherPaymentMethodDetails {
  type: Type;
  name: Name;
}
export interface CardPaymentMethodDetails {
  type: Type1;
  name: Name1;
  brand?: Brand;
  firstDigits?: Firstdigits;
  lastDigits?: Lastdigits;
  expMonth?: Expmonth;
  expYear?: Expyear;
}
export interface TransactionSessionFailure {
  pspReference?: Pspreference1;
  amount?: Amount1;
  time?: Time1;
  externalUrl?: Externalurl1;
  message?: Message1;
  actions?: Actions1;
  result: Result1;
  data?: JsonValue1;
  paymentMethodDetails?: Paymentmethoddetails1;
}
export interface TransactionSessionActionRequired {
  pspReference?: Pspreference2;
  amount?: Amount2;
  time?: Time2;
  externalUrl?: Externalurl2;
  message?: Message2;
  actions?: Actions2;
  result: Result2;
  data?: JsonValue2;
  paymentMethodDetails?: Paymentmethoddetails2;
}